fix(sgp): register as.double method so S3 dispatch works for coercion
as.numeric() is a base primitive that does not dispatch S3 methods for
list objects; as.double() does. Renamed as.numeric.sgp_denominators to
as.double.sgp_denominators so that both as.double(x) and as.numeric(x)
(which internally calls as.double) correctly return the denominator vector
rather than falling back to base R and failing on a list object. Updated
roxygen @method tag and regenerated NAMESPACE via devtools::document().

+++ R/sgp-denominators-s3.R
@@ -81,14 +81,19 @@ print.sgp_denominators <- function(x, ...) {
 #' Coerce sgp_denominators to a named numeric vector
 #'
 #' Returns the underlying denominator vector. Enables backward-compatible
-#' usage such as `as.numeric(denoms)`.
+#' usage such as `as.double(denoms)` or `as.numeric(denoms)`.
+#'
+#' `as.numeric()` is a base primitive that does not dispatch S3 methods for
+#' list objects; `as.double()` does. Registering as `as.double` ensures
+#' correct dispatch for both `as.double()` and `as.numeric()` callers (since
+#' `as.numeric` internally calls `as.double`).
 #'
 #' @param x An `sgp_denominators` object.
 #' @param ... Ignored.
 #' @return Named numeric vector of denominators.
-#' @method as.numeric sgp_denominators
+#' @method as.double sgp_denominators
 #' @export
-as.numeric.sgp_denominators <- function(x, ...) x$denominators
+as.double.sgp_denominators <- function(x, ...) x$denominators
 
 #' Names of an sgp_denominators object
 #'
